save user related personal information data in mocks
