+++ frontend/app/.server/domain/services/profile-service-mock.ts
@@ -1,12 +1,6 @@
 import { Err, Ok } from 'oxide.ts';
 import type { Result, Option } from 'oxide.ts';
 
-import { getCityService } from './city-service';
-import { getClassificationService } from './classification-service';
-import { getDirectorateService } from './directorate-service';
-import { getEmploymentOpportunityTypeService } from './employment-opportunity-type-service';
-import { getWFAStatuses } from './wfa-status-service';
-
 import type {
   Profile,
   ProfilePutModel,
@@ -15,9 +9,15 @@ import type {
   CollectionProfileResponse,
   ProfileQueryParams,
 } from '~/.server/domain/models';
+import { getCityService } from '~/.server/domain/services/city-service';
+import { getClassificationService } from '~/.server/domain/services/classification-service';
+import { getDirectorateService } from '~/.server/domain/services/directorate-service';
+import { getEmploymentOpportunityTypeService } from '~/.server/domain/services/employment-opportunity-type-service';
 import { getLanguageForCorrespondenceService } from '~/.server/domain/services/language-for-correspondence-service';
 import { getLanguageReferralTypeService } from '~/.server/domain/services/language-referral-type-service';
+import { createAndLinkNewMockProfile, mockProfiles } from '~/.server/domain/services/mockData';
 import type { ProfileService } from '~/.server/domain/services/profile-service';
+import { getWFAStatuses } from '~/.server/domain/services/wfa-status-service';
 import { PROFILE_STATUS_ID } from '~/domain/constants';
 import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
@@ -198,7 +198,7 @@ export function getMockProfileService(): ProfileService {
     async registerProfile(accessToken: string): Promise<Result<Profile, AppError>> {
       debugLog('registerProfile', 'Attempting to register new profile', { accessTokenLength: accessToken.length });
 
-      const newProfile = createMockProfile(accessToken);
+      const newProfile = createAndLinkNewMockProfile(accessToken);
       debugLog('registerProfile', 'Successfully registered new profile', {
         profileId: newProfile.id,
         profileStatus: newProfile.profileStatus?.code,
@@ -412,361 +412,4 @@ export function getMockProfileService(): ProfileService {
       return result.ok();
     },
   };
-}
-
-/**
- * Mock profile data for testing and development.
- */
-const mockProfiles: Profile[] = [
-  {
-    id: 1,
-    additionalComment: 'Looking for opportunities in software development.',
-    hasConsentedToPrivacyTerms: false,
-    hrAdvisorId: undefined,
-    isAvailableForReferral: undefined,
-    isInterestedInAlternation: undefined,
-    personalEmailAddress: 'jane.doe@example.com',
-    personalPhoneNumber: '613-555-0001',
-    languageOfCorrespondence: {
-      id: 0,
-      code: 'EN',
-      nameEn: 'English',
-      nameFr: 'Anglais',
-    },
-    profileStatus: {
-      id: 2,
-      code: 'INCOMPLETE',
-      nameEn: 'Incomplete',
-      nameFr: 'Incomplet',
-    },
-    profileUser: {
-      id: 1,
-      businessEmailAddress: 'jane.doe@canada.ca',
-      businessPhoneNumber: '+1-613-555-0101',
-      firstName: 'Jane',
-      initial: 'D',
-      lastName: 'Doe',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000001',
-      personalRecordIdentifier: '123456789',
-      language: {
-        id: 1,
-        code: 'EN',
-        nameEn: 'English',
-        nameFr: 'Anglais',
-      },
-      userType: {
-        id: 3,
-        code: 'HRA',
-        nameEn: 'HR Advisor',
-        nameFr: 'Conseiller RH',
-      },
-      createdBy: 'system',
-      createdDate: '2024-01-01T00:00:00Z',
-      lastModifiedBy: 'system',
-      lastModifiedDate: '2024-01-01T00:00:00Z',
-    },
-    createdBy: 'system',
-    createdDate: '2024-01-01T00:00:00Z',
-    lastModifiedBy: undefined,
-    lastModifiedDate: undefined,
-  },
-  {
-    id: 2,
-    additionalComment: undefined,
-    hasConsentedToPrivacyTerms: false,
-    hrAdvisorId: undefined,
-    isAvailableForReferral: undefined,
-    isInterestedInAlternation: undefined,
-    personalEmailAddress: undefined,
-    personalPhoneNumber: undefined,
-    languageOfCorrespondence: undefined,
-    profileStatus: {
-      id: 2,
-      code: 'INCOMPLETE',
-      nameEn: 'Incomplete',
-      nameFr: 'Incomplet',
-    },
-    profileUser: {
-      id: 2,
-      businessEmailAddress: 'john.smith@canada.ca',
-      businessPhoneNumber: '+1-613-555-0102',
-      firstName: 'John',
-      initial: 'S',
-      lastName: 'Smith',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000002',
-      personalRecordIdentifier: '987654321',
-      language: {
-        id: 1,
-        code: 'EN',
-        nameEn: 'English',
-        nameFr: 'Anglais',
-      },
-      userType: {
-        id: 1,
-        code: 'employee',
-        nameEn: 'Employee',
-        nameFr: 'Employé',
-      },
-      createdBy: 'system',
-      createdDate: '2024-01-01T00:00:00Z',
-      lastModifiedBy: 'jane.doe',
-      lastModifiedDate: '2024-01-15T10:30:00Z',
-    },
-    createdBy: 'system',
-    createdDate: '2024-01-01T00:00:00Z',
-    lastModifiedBy: 'jane.doe',
-    lastModifiedDate: '2024-01-15T10:30:00Z',
-  },
-  {
-    id: 3,
-    additionalComment: 'Interested in remote work.',
-    hasConsentedToPrivacyTerms: true,
-    hrAdvisorId: 1,
-    isAvailableForReferral: false,
-    isInterestedInAlternation: true,
-    personalEmailAddress: 'john.smith@example.com',
-    personalPhoneNumber: '613-555-0001',
-    languageOfCorrespondence: {
-      id: 1,
-      code: 'EN',
-      nameEn: 'English',
-      nameFr: 'Anglais',
-    },
-    profileStatus: {
-      id: 0,
-      code: 'PENDING',
-      nameEn: 'Pending',
-      nameFr: 'En attente',
-    },
-    profileUser: {
-      id: 3,
-      businessEmailAddress: 'john.smith@canada.ca',
-      businessPhoneNumber: '+1-613-555-0103',
-      firstName: 'John',
-      initial: 'S',
-      lastName: 'Smith',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000003',
-      personalRecordIdentifier: '987654321',
-      language: {
-        id: 1,
-        code: 'EN',
-        nameEn: 'English',
-        nameFr: 'Anglais',
-      },
-      userType: {
-        id: 1,
-        code: 'employee',
-        nameEn: 'Employee',
-        nameFr: 'Employé',
-      },
-      createdBy: 'system',
-      createdDate: '2024-02-01T09:00:00Z',
-      lastModifiedBy: 'john.smith',
-      lastModifiedDate: '2024-02-10T14:00:00Z',
-    },
-    createdBy: 'system',
-    createdDate: '2024-02-01T09:00:00Z',
-    lastModifiedBy: 'john.smith',
-    lastModifiedDate: '2024-02-10T14:00:00Z',
-  },
-  {
-    id: 4,
-    additionalComment: 'Fluent in French and English.',
-    hasConsentedToPrivacyTerms: true,
-    hrAdvisorId: 1,
-    isAvailableForReferral: true,
-    isInterestedInAlternation: false,
-    personalEmailAddress: 'marie.curie@example.com',
-    personalPhoneNumber: '613-555-5555',
-    languageOfCorrespondence: {
-      id: 2,
-      code: 'FR',
-      nameEn: 'French',
-      nameFr: 'Français',
-    },
-    profileStatus: {
-      id: 1,
-      code: 'APPROVED',
-      nameEn: 'Approved',
-      nameFr: 'Approuvé',
-    },
-    profileUser: {
-      id: 4,
-      businessEmailAddress: 'marie.curie@canada.ca',
-      businessPhoneNumber: '+1-613-555-0104',
-      firstName: 'Marie',
-      initial: 'C',
-      lastName: 'Curie',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000004',
-      personalRecordIdentifier: '555555555',
-      language: {
-        id: 2,
-        code: 'FR',
-        nameEn: 'French',
-        nameFr: 'Français',
-      },
-      userType: {
-        id: 1,
-        code: 'employee',
-        nameEn: 'Employee',
-        nameFr: 'Employé',
-      },
-      createdBy: 'system',
-      createdDate: '2024-03-15T08:30:00Z',
-      lastModifiedBy: 'john.smith',
-      lastModifiedDate: '2024-02-10T14:00:00Z',
-    },
-    createdBy: 'system',
-    createdDate: '2024-03-15T08:30:00Z',
-    lastModifiedBy: 'john.smith',
-    lastModifiedDate: '2024-02-10T14:00:00Z',
-  },
-  {
-    id: 5,
-    additionalComment: 'Open to contract roles.',
-    hasConsentedToPrivacyTerms: true,
-    hrAdvisorId: 5,
-    isAvailableForReferral: true,
-    isInterestedInAlternation: true,
-    personalEmailAddress: 'alex.tan@example.com',
-    personalPhoneNumber: '613-555-3333',
-    languageOfCorrespondence: {
-      id: 1,
-      code: 'EN',
-      nameEn: 'English',
-      nameFr: 'Anglais',
-    },
-    profileStatus: {
-      id: 1,
-      code: 'APPROVED',
-      nameEn: 'Approved',
-      nameFr: 'Approuvé',
-    },
-    profileUser: {
-      id: 5,
-      businessEmailAddress: 'alex.tan@canada.ca',
-      businessPhoneNumber: '+1-613-555-0105',
-      firstName: 'Alex',
-      initial: 'T',
-      lastName: 'Tan',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000005',
-      personalRecordIdentifier: '222333444',
-      language: {
-        id: 1,
-        code: 'EN',
-        nameEn: 'English',
-        nameFr: 'Anglais',
-      },
-      userType: {
-        id: 1,
-        code: 'employee',
-        nameEn: 'Employee',
-        nameFr: 'Employé',
-      },
-      createdBy: 'system',
-      createdDate: '2024-04-20T11:45:00Z',
-      lastModifiedBy: 'alex.tan',
-      lastModifiedDate: '2024-05-01T10:00:00Z',
-    },
-    createdBy: 'system',
-    createdDate: '2024-04-20T11:45:00Z',
-    lastModifiedBy: 'alex.tan',
-    lastModifiedDate: '2024-05-01T10:00:00Z',
-  },
-];
-
-// Mock Active Directory ID to User ID mapping for profiles
-const activeDirectoryToUserIdMap: Record<string, number> = {
-  '00000000-0000-0000-0000-000000000001': 1,
-  '11111111-1111-1111-1111-111111111111': 2,
-};
-
-/**
- * A private helper function to generate a mock profile.
- *
- * @param accessToken The access token of the user to create a profile for.
- * @returns The created profile object.
- * @throws {AppError} If the profile cannot be created (e.g., user not found).
- */
-function createMockProfile(accessToken: string): Profile {
-  debugLog('createMockProfile', 'Creating new mock profile', { accessTokenLength: accessToken.length });
-  let userId = activeDirectoryToUserIdMap[accessToken];
-  if (!userId) {
-    // Create new entry in activeDirectoryToUserIdMap if it doesn't exist
-    userId = mockProfiles.length + 1;
-    activeDirectoryToUserIdMap['00000000-0000-0000-0000-000000000000'] = userId;
-    debugLog('createMockProfile', `Created new user mapping: userId ${userId}`);
-  } else {
-    debugLog('createMockProfile', `Using existing user mapping: userId ${userId}`);
-  }
-
-  // Create new profile
-  const newProfile: Profile = {
-    id: mockProfiles.length + 1,
-    additionalComment: 'Looking for opportunities in software development.',
-    hasConsentedToPrivacyTerms: false,
-    hrAdvisorId: undefined,
-    isAvailableForReferral: true,
-    isInterestedInAlternation: false,
-    personalEmailAddress: 'personal.email@example.com',
-    personalPhoneNumber: '613-938-0001',
-    languageOfCorrespondence: {
-      id: 1,
-      code: 'EN',
-      nameEn: 'English',
-      nameFr: 'Anglais',
-    },
-    profileStatus: {
-      id: 2,
-      code: 'INCOMPLETE',
-      nameEn: 'Incomplete',
-      nameFr: 'Incomplet',
-    },
-    profileUser: {
-      id: userId,
-      businessEmailAddress: 'work.email@example.ca',
-      businessPhoneNumber: undefined,
-      firstName: 'John',
-      initial: 'D',
-      lastName: 'Doe',
-      middleName: undefined,
-      microsoftEntraId: '00000000-0000-0000-0000-000000000000',
-      personalRecordIdentifier: '123456789',
-      language: {
-        id: 1,
-        code: 'EN',
-        nameEn: 'English',
-        nameFr: 'Anglais',
-      },
-      userType: {
-        id: 1,
-        code: 'employee',
-        nameEn: 'Employee',
-        nameFr: 'Employé',
-      },
-      createdBy: 'system',
-      createdDate: new Date().toISOString(),
-      lastModifiedBy: 'jane.doe',
-      lastModifiedDate: new Date().toISOString(),
-    },
-    createdBy: accessToken,
-    createdDate: new Date().toISOString(),
-    lastModifiedBy: accessToken,
-    lastModifiedDate: new Date().toISOString(),
-  };
-
-  // Add the new profile to the mock profiles array
-  mockProfiles.push(newProfile);
-
-  debugLog('createMockProfile', 'Successfully created mock profile', {
-    profileId: newProfile.id,
-    userId,
-    totalProfiles: mockProfiles.length,
-  });
-  return newProfile;
 }
